+++ php/Terminus/Auth.php
@@ -20,28 +20,48 @@ class Auth {
   }
 
   /**
-   * Determines if user is logged in
+   * Ensures the user is logged in or errs.
    *
-   * @return [boolean] True if user is logged in
+   * @return [boolean] Always true
    */
-  public static function loggedIn() {
-    if (Session::instance()->getValue('session', false) === false) {
-      throw new TerminusException(
-        'Please login first with `terminus auth login`',
-        array(),
-        1
-      );
+  public static function ensureLogin() {
+    $session = Session::instance()->getData();
+    $auth    = new Auth();
+    if (!$auth->loggedIn()) {
+      if (isset($session->refresh)) {
+        $auth->logInViaMachineToken($session->refresh);
+      } else {
+        throw new TerminusException(
+          'Please login first with `terminus auth login`',
+          array(),
+          1
+        );
+      }
     }
     return true;
   }
 
   /**
-   * Execute the login based on an existing session token
+   * Checks to see if the current user is logged in
    *
-   * @param [string] $token Session token to initiate login with
+   * @return [boolean] $is_logged_in True if the user is logged in
+   */
+  public function loggedIn() {
+    $session = Session::instance()->getData();
+    $is_logged_in = (
+      isset($session->session)
+      && (Terminus::isTest() || ($session->session_expire_time >= time()))
+    );
+    return $is_logged_in;
+  }
+
+  /**
+   * Execute the login based on a machine token
+   *
+   * @param [string] $token Machine token to initiate login with
    * @return [boolean] True if login succeeded
    */
-  public function logInViaRefreshToken($token) {
+  public function logInViaMachineToken($token) {
     $options = array(
       'headers' => array('Content-type' => 'application/json'),
       'body'    => array(
@@ -49,7 +69,7 @@ class Auth {
       ),
     );
 
-    $this->logger->info('Logging in via refresh token');
+    $this->logger->info('Logging in via machine token');
     $response = TerminusCommand::request('auth/refresh', '', '', 'POST', $options);
 
     if (!$response
@@ -57,8 +77,8 @@ class Auth {
       || $response['info']['http_code'] != '200'
     ) {
       throw new TerminusException(
-        'The refresh token {token} is not valid.',
-        compact('token'),
+        'The provided machine token is not valid.',
+        array(),
         1
       );
     }
@@ -66,7 +86,8 @@ class Auth {
       'Logged in as {uuid}.',
       array('uuid' => $response['data']->user_id)
     );
-
+    $data = $response['data'];
+    $data->refresh = $token;
     $this->setInstanceData($response['data']);
     return true;
   }
@@ -115,15 +136,23 @@ class Auth {
   /**
    * Saves the session data to a cookie
    *
-   * @param [array] $session Session data to save
+   * @param [array] $data Session data to save
    * @return [boolean] Always true
    */
   private function setInstanceData($data) {
+    if (!isset($data->refresh)) {
+      $refresh = (array)Session::instance()->get('refresh');
+    } else {
+      $refresh = $data->refresh;
+    }
     $session = array(
       'user_uuid'           => $data->user_id,
       'session'             => $data->session,
       'session_expire_time' => $data->expires_at,
     );
+    if ($refresh && is_string($refresh)) {
+      $session['refresh'] = $refresh;
+    }
     Session::instance()->setData($session);
     return true;
   }